+++ vault/provider.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"fmt"
 	"log"
+	"os"
 	"strings"
 
 	"github.com/hashicorp/go-multierror"
@@ -49,6 +50,12 @@ func Provider() terraform.ResourceProvider {
 				Required:    true,
 				DefaultFunc: schema.EnvDefaultFunc("VAULT_ADDR", nil),
 				Description: "URL of the root of the target Vault server.",
+			},
+			"add_address_to_env": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Default:     false,
+				Description: "If true, adds the value of the `address` argument to the Terraform process environment.",
 			},
 			"token": {
 				Type:        schema.TypeString,
@@ -203,6 +210,10 @@ var (
 		"vault_aws_access_credentials": {
 			Resource:      awsAccessCredentialsDataSource(),
 			PathInventory: []string{"/aws/creds"},
+		},
+		"vault_azure_access_credentials": {
+			Resource:      azureAccessCredentialsDataSource(),
+			PathInventory: []string{"/azure/creds/{role}"},
 		},
 		"vault_generic_secret": {
 			Resource:      genericSecretDataSource(),
@@ -564,6 +575,22 @@ func providerToken(d *schema.ResourceData) (string, error) {
 	if token := d.Get("token").(string); token != "" {
 		return token, nil
 	}
+
+	if addAddr := d.Get("add_address_to_env").(string); addAddr == "true" {
+		if addr := d.Get("address").(string); addr != "" {
+			if current, exists := os.LookupEnv("VAULT_ADDR"); exists {
+				defer func() {
+					os.Setenv("VAULT_ADDR", current)
+				}()
+			} else {
+				defer func() {
+					os.Unsetenv("VAULT_ADDR")
+				}()
+			}
+			os.Setenv("VAULT_ADDR", addr)
+		}
+	}
+
 	// Use ~/.vault-token, or the configured token helper.
 	tokenHelper, err := config.DefaultTokenHelper()
 	if err != nil {